move attrs compat to utils
